requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
tqdm==4.66.4
librosa== 0.11.0
numba==0.58.1
numpy==2.2.1
imutils==0.5.4
onnx==1.17.0
onnxruntime-gpu==1.20.1
opencv-contrib-python==4.9.0.80